pyproject.toml
[project]
name = "rd-tablebench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs>=2.4.4",
    "aiohttp>=3.11.11",
    "aiolimiter>=1.2.1",
    "aiosignal>=1.3.2",
    "annotated-types>=0.7.0",
    "anthropic>=0.43.0",
    "anyio>=4.8.0",
    "attrs==24.3.0",
    "backoff>=2.2.1",
    "cachetools>=5.5.0",
    "certifi>=2024.12.14",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "distro>=1.9.0",
    "docstring-parser>=0.16",
    "frozenlist>=1.5.0",
    "google-ai-generativelanguage>=0.6.10",
    "google-api-core==2.24.0",
    "google-api-python-client>=2.158.0",
    "google-auth>=2.37.0",
    "google-auth-httplib2>=0.2.0",
    "google-cloud-aiplatform==1.76.0",
    "google-cloud-bigquery==3.27.0",
    "google-cloud-core==2.4.1",
    "google-cloud-resource-manager==1.14.0",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.6.0",
    "google-generativeai==0.8.3",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.66.0",
    "grpc-google-iam-v1==0.14.0",
    "grpcio==1.69.0",
    "grpcio-status==1.69.0",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.8.2",
    "jsonref==1.1.0",
    "levenshtein==0.26.1",
    "lxml==5.3.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "modal>=1.2.2",
    "multidict==6.1.0",
    "packaging==24.2",
    "pdf2image==1.17.0",
    "pillow>=11.1.0",
    "polars==1.19.0",
    "propcache==0.2.1",
    "proto-plus==1.25.0",
    "protobuf>=5.29.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic>=2.10.5",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pymupdf>=1.24.0",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "rapidfuzz==3.11.0",
    "requests==2.32.3",
    "rich==13.9.4",
    "rsa==4.9",
    "shapely==2.0.6",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sycamore-ai",
    "tqdm==4.67.1",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "uritemplate==4.1.1",
    "urllib3>=2.3.0",
    "yarl==1.18.3",
]

[tool.uv.sources]
sycamore-ai = { path = "../sycamore/lib/sycamore/", editable = true }

[dependency-groups]
dev = [
    "lxml-stubs>=0.5.1",
    "types-requests>=2.32.4.20250913",
    "types-tqdm>=4.67.0.20250809",
]
